requirements.txt
numpy
rouge_score
fire
openai
async-openai
transformers
torch
sentencepiece
tokenizers==0.12.1
#wandb
#accelerate

requests
beautifulsoup4
tabulate
youtube_search

fastcore
inflect
ipywidgets
send2trash
ipynbname
nbdev

argh
sh
lxml
pytest
spacy

pydub
SpeechRecognition
openai-whisper
sounddevice
soundfile
gTTS
TTS
clip-interrogator

watchdog

anthropic
aionotify

starlette
uvicorn
watchfiles

inotify

websockets
pyaudio

markdown
aiofiles
jinja2
python-multipart

pyjwt
supabase_py
markdown-katex

segment-anything

tiktoken

pyalsaaudio
wordpress_api
selenium
